refactor(test): move responses to helper
