+++ test/__init__.py
@@ -1,3 +1,3 @@
-from .helper_functions import encode_response
+from .helper import encode_response
 
 __all__ = ['encode_response']
