--- test/helper_functions.py
@@ -1,5 +1,0 @@
-import json
-
-
-def encode_response(expected_data):
-    return json.dumps(expected_data).encode('utf-8')
